backend/integration/requirements.txt
affine==2.3.0; python_version >= "3.9" and python_version < "3.10"
args==0.1.0; python_version >= "3.9" and python_version < "3.10"
attrs==21.2.0; python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.5.0"
certifi==2021.5.30; python_version >= "3.9" and python_version < "3.10" and (python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.6.0")
charset-normalizer==2.0.4; python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.6.0"
./ci/dist/ci-0.1.6-py3-none-any.whl
click-plugins==1.1.1; python_version >= "3.9" and python_version < "3.10"
click==8.0.1; python_version >= "3.9" and python_version < "3.10" and (python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_full_version >= "3.3.0" and python_version < "3.10" and python_version >= "3.9")
cligj==0.7.2; python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_full_version >= "3.3.0" and python_version < "3.10" and python_version >= "3.9"
clint==0.5.1; python_version >= "3.9" and python_version < "3.10"
colorama==0.4.4; python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" and platform_system == "Windows" or python_version >= "3.9" and python_version < "3.10" and platform_system == "Windows" and python_full_version >= "3.5.0"
coverage==5.5; python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_full_version >= "3.5.0" and python_version < "3.10" and python_version >= "3.9"
cycler==0.10.0; python_version >= "3.9" and python_version < "3.10"
et-xmlfile==1.1.0; python_version >= "3.6"
fiona==1.8.20; python_version >= "3.9" and python_version < "3.10"
gdal==3.2.2; python_version >= "3.9" and python_version < "3.10"
geopandas==0.9.0; python_version >= "3.9" and python_version < "3.10"
idna==3.2; python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.6.0"
kiwisolver==1.3.1; python_version >= "3.9" and python_version < "3.10"
matplotlib==3.4.2; python_version >= "3.9" and python_version < "3.10"
munch==2.5.0; python_version >= "3.9" and python_version < "3.10"
numpy==1.21.1; python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.7.1"
openpyxl==3.0.7; python_version >= "3.6"
pandas==1.3.1; python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.7.1"
pillow==8.3.1; python_version >= "3.9" and python_version < "3.10"
./power/dist/power-0.2.0-py3-none-any.whl
pycrs==1.0.2; python_version >= "3.9" and python_version < "3.10"
pydantic==1.8.2; python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.6.1"
pyparsing==2.4.7; python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.3.0"
pyproj==3.1.0; python_version >= "3.9" and python_version < "3.10"
pyshp==2.1.3; python_version >= "3.9" and python_version < "3.10"
pysimplegui==4.70.1; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2; python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.7.1" and (python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.3.0")
pytz==2021.1; python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.7.1"
pyyaml==5.4.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
rasterio==1.2.6; python_version >= "3.9" and python_version < "3.10"
rasterstats==0.15.0; python_version >= "3.9" and python_version < "3.10"
requests==2.26.0; python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.6.0"
scipy==1.7.1; python_version >= "3.9" and python_version < "3.10"
shapely==1.7.1; python_version >= "3.9" and python_version < "3.10"
simplejson==3.17.3; python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.3.0"
six==1.16.0; python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.7.1" and (python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.3.0")
snuggs==1.4.7; python_version >= "3.9" and python_version < "3.10"
typing-extensions==3.10.0.0; python_version >= "3.9" and python_version < "3.10" and python_full_version >= "3.6.1"
./ubm/dist/ubm-0.9.0-py3-none-any.whl
urllib3==1.26.6; python_version >= "3.9" and python_full_version < "3.0.0" and python_version < "3.10" or python_full_version >= "3.6.0" and python_version < "3.10" and python_version >= "3.9"
xlsxwriter==1.4.5
jwt==1.3.1
pycryptodome==3.15.0
./ci-and-p-gui/dist/gui-0.2.2-py3-none-any.whl
pymongo